Using LLaMa for benchmark, `python3 -O` to disable asserts as some of them are expensive and we don't want to sacrifice readability by optimizing them

```
GPU=1 DEBUG=0 OPTLOCAL=2 python3 -O examples/llama.py --prompt "Hello." --count 10 --temperature 0 --timing --profile
```

Aggregating by time spent per file

Before:

```
lazy.py: 0.9459 seconds
shapetracker.py: 0.8728 seconds
tensor.py: 0.8376 seconds
ops.py: 0.2657 seconds
symbolic.py: 0.1064 seconds
```

After:

```
lazy.py: 0.9416 seconds
tensor.py: 0.8587 seconds
shapetracker.py: 0.4083 seconds
ops.py: 0.3439 seconds
symbolic.py: 0.1100 seconds
```

Saving 0.47s on `shapetracker.py`