@tinyb0t 2023-06-10
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/shape/shapetracker.py           103              65
------------------------------------------------------------
lines added in the tinygrad folder: 38
```