@geohot 2023-06-10
Overall looks pretty good.

I bet the core gains come from breaking out and caching the helper functions, I think that's a win for both speed and readability.

I think that anywhere we are trading off readability for minor speed it's not worth it. I'm curious to see microbenchmarks on:

`isinstance vs __class__ ==`
`tuple(gen) vs tuple([gen])` (this is a bug in python I think if right is faster)
`[x for x in lst] vs map(fxn, lst)`

I generally find the left side more readable.