@rayanht 2023-06-10
`x[::-1]` vs. `reversed` 

X = [1, 2, 3, 4, 5, 6, 7, 8]

```
def f1(x):
    return sum(reversed(x))

def f2(x):
    return sum(x[::-1])

t1 = timeit.timeit(lambda: f1(X), number=1000000)
t2 = timeit.timeit(lambda: f2(X), number=1000000)

print("f1 took %f seconds" % t1)
print("f2 took %f seconds" % t2)
```

f1 took 0.1166 seconds
f2 took 0.1260 seconds

[::-1] creates a copy whilst reverse is just an iterator on top of the same underlying data, I feel like it's also more readable