@roelofvandijk 2023-06-11
I think you really want to use `is` for the nodes, ops and type checking.

```
import timeit
from tinygrad.shape.symbolic import NumNode
X = [NumNode(x) for x in range(100)]

def f1(x):
    return [isinstance(x, NumNode) for x in X]
def f2(x):
    return [x.__class__ == NumNode for x in X]

def f3(x):
    return [x.__class__ is NumNode for x in X]

t1 = timeit.timeit(lambda: f1(X), number=1000000)
t2 = timeit.timeit(lambda: f2(X), number=1000000)
t3 = timeit.timeit(lambda: f3(X), number=1000000)

print("f1 took %f seconds" % t1)
print("f2 took %f seconds" % t2)
print("f3 took %f seconds" % t3)
```
f1 took 2.744003 seconds
f2 took 2.677994 seconds
f3 took 2.071155 seconds