@geohot 2023-06-11
Cool, okay so:
* `reversed` is good, small gain but no real readability hit
* list in tuple is good, pretty big gain for very minor readability hit
* for the 3% loss I'd say isinstance is worth it, but using is instead of equals is a 25% gain, so let's do that