@roelofvandijk 2023-06-11
```
import timeit
X = list(range(100))
X_range = range(100)

def func(x): return x**7
def f1(lst): return [func(x) for x in lst]
def f2(lst): return list((func(x) for x in lst))
def f3(lst): return list(map(func, lst))

t1 = timeit.timeit(lambda: f1(X), number=100000)
t2 = timeit.timeit(lambda: f2(X), number=100000)
t3 = timeit.timeit(lambda: f3(X), number=100000)

print("%f seconds: f1 (list comprehension on list)" % t1)
print("%f seconds: f2 (generator comprehension on list)" % t2)
print("%f seconds: f3 (map on list)" % t3)

t4 = timeit.timeit(lambda: f1(X_range), number=100000)
t5 = timeit.timeit(lambda: f2(X_range), number=100000)
t6 = timeit.timeit(lambda: f3(X_range), number=100000)

print("%f seconds: f1 (list comprehension on generator)" % t4)
print("%f seconds: f2 (generator comprehension on generator)" % t5)
print("%f seconds: f3 (map on generator)" % t6)
```
0.967163 seconds: f1 (list comprehension on list)
1.124679 seconds: f2 (generator comprehension on list)
0.896665 seconds: f3 (map on list)
0.938883 seconds: f1 (list comprehension on generator)
1.076224 seconds: f2 (generator comprehension on generator)
0.895292 seconds: f3 (map on generator)