@geohot 2023-06-11
How does that compare to `[x**3 for x in lst]`?