@roelofvandijk 2023-06-11
Quite unfavourably (function call overhead).

```
import timeit

def func(x): return x**3
def f1(lst): return [func(x) for x in lst]
def f2(lst): return list((func(x) for x in lst))
def f3(lst): return list(map(func, lst))
def f4(lst): return [x**3 for x in X_range]

for n in [5,1000]:
  X = list(range(n))
  t1 = timeit.timeit(lambda: f1(X), number=100000)
  t2 = timeit.timeit(lambda: f2(X), number=100000)
  t3 = timeit.timeit(lambda: f3(X), number=100000)

  X_range = range(n)
  t4 = timeit.timeit(lambda: f1(X_range), number=100000)
  t5 = timeit.timeit(lambda: f2(X_range), number=100000)
  t6 = timeit.timeit(lambda: f3(X_range), number=100000)

  t7 = timeit.timeit(lambda: f3(X), number=100000)
  t8 = timeit.timeit(lambda: f4(X_range), number=100000)

  print(f"Number of elements: {n}")
  print("%f seconds: f1 (list comprehension on list)" % t1)
  print("%f seconds: f2 (generator comprehension on list)" % t2)
  print("%f seconds: f3 (map on list)" % t3)
  print("%f seconds: f1 (list comprehension on generator)" % t4)
  print("%f seconds: f2 (generator comprehension on generator)" % t5)
  print("%f seconds: f3 (map on generator)" % t6)

  print("%f seconds: direct list comprehension on list" % t7)
  print("%f seconds: direct list comprehension on generator" % t8)
```
Number of elements: 5
0.044887 seconds: f1 (list comprehension on list)
0.056025 seconds: f2 (generator comprehension on list)
0.041223 seconds: f3 (map on list)
0.048437 seconds: f1 (list comprehension on generator)
0.055826 seconds: f2 (generator comprehension on generator)
0.042119 seconds: f3 (map on generator)
0.039749 seconds: direct list comprehension on list
0.030753 seconds: direct list comprehension on generator
Number of elements: 1000
6.249403 seconds: f1 (list comprehension on list)
7.280469 seconds: f2 (generator comprehension on list)
5.923099 seconds: f3 (map on list)
6.903187 seconds: f1 (list comprehension on generator)
7.844344 seconds: f2 (generator comprehension on generator)
6.360970 seconds: f3 (map on generator)
5.836845 seconds: direct list comprehension on list
4.585988 seconds: direct list comprehension on generator

That's why `def prod(self, x): return math.prod(x)`  and `def shape(self): return self.views[-1].shape` were slow.