@rayanht 2023-06-12
Updated the map to a list comp

@roelofvandijk I'd rather not micro-optimize the existing assertions in the scope of this PR, we're already changing a lot of moving parts and a real performance run would use the `-O` flag anyway to skip them