@geohot 2023-06-13
I'm going to merge this. I think there's a few more `==` to change to `is` and an extra `return self`, but it's better than what is there.