@geohot 2023-06-13
203.27 ms -> 171.39 ms on my M1 Max